PR Feedback.

+++ libs/unity/library/Runtime/Shaders/YUVFeedShaderUnlit.shader
@@ -55,22 +55,22 @@ Shader "Microsoft.MixedReality.WebRTC/YUVFeedShader (unlit)"
 			sampler2D _UPlane;
 			sampler2D _VPlane;
 
-			float3 yuv2rgb(float3 yuv)
+			half3 yuv2rgb(float3 yuv)
 			{
 				// The YUV to RBA conversion, please refer to: http://en.wikipedia.org/wiki/YUV
 				// Y'UV420p (I420) to RGB888 conversion section.
-				float y_value = yuv[0];
-				float u_value = yuv[1];
-				float v_value = yuv[2];
-				float r = y_value + 1.370705 * (v_value - 0.5);
-				float g = y_value - 0.698001 * (v_value - 0.5) - (0.337633 * (u_value - 0.5));
-				float b = y_value + 1.732446 * (u_value - 0.5);
-				return float3(r, g, b);
+				half y_value = yuv[0];
+				half u_value = yuv[1];
+				half v_value = yuv[2];
+				half r = y_value + 1.370705 * (v_value - 0.5);
+				half g = y_value - 0.698001 * (v_value - 0.5) - (0.337633 * (u_value - 0.5));
+				half b = y_value + 1.732446 * (u_value - 0.5);
+				return half3(r, g, b);
 			}
 
 			fixed4 frag (v2f i) : SV_Target
 			{
-				float3 yuv;
+				half3 yuv;
 				yuv.x = tex2D(_YPlane, i.uv).r;
 				yuv.y = tex2D(_UPlane, i.uv).r;
 				yuv.z = tex2D(_VPlane, i.uv).r;
